fix(web): keep diff palette mapping in sync
